modified tests for binac

+++ binac/cupy/example7.py
@@ -3,6 +3,6 @@ import cupyx.scipy.fft as cufft
 import scipy.fft
 scipy.fft.set_global_backend(cufft)
 
-a = cp.random.random(100).astype(cp.complex64)
+a = cp.random.random(100).astype(cp.float64)
 b = scipy.fft.fft(a)
 print('Success')
